// src/lib/okane-finance-api/generated/models/DominantRegime.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const DominantRegime = {
    Bull: 'bull',
    Bear: 'bear',
    Chop: 'chop'
} as const;
export type DominantRegime = typeof DominantRegime[keyof typeof DominantRegime];


export function DominantRegimeFromJSON(json: any): DominantRegime {
    return DominantRegimeFromJSONTyped(json, false);
}

export function DominantRegimeFromJSONTyped(json: any, ignoreDiscriminator: boolean): DominantRegime {
    return json as DominantRegime;
}

export function DominantRegimeToJSON(value?: DominantRegime | null): any {
    return value as any;
}